+++ src/server/components/features/features-list/NewTablePluginForWizardAndQl.ts
@@ -4,7 +4,7 @@ import {createFeatureConfig} from '../utils';
 export default createFeatureConfig({
     name: Feature.NewTablePluginForWizardAndQl,
     state: {
-        development: false,
-        production: false,
+        development: true,
+        production: true,
     },
 });
