+++ src/server/components/features/features-list/PinnedColumns.ts
@@ -4,7 +4,7 @@ import {createFeatureConfig} from '../utils';
 export default createFeatureConfig({
     name: Feature.PinnedColumns,
     state: {
-        development: false,
-        production: false,
+        development: true,
+        production: true,
     },
 });
