+++ src/ui/components/Table/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useDevicePixelRatio';
+export * from './useTableDimensions';
