+++ src/ui/components/Table/hooks/useDevicePixelRatio.ts
