+++ src/ui/libs/DatalensChartkit/ChartKit/plugins/Table/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useTableEvents';
